Use the BouncyCastleProvider class directly (#84)
* Use the BouncyCastleProvider class directly

This avoids issues where a shaded version of BCProvider is installed in the
system JCE provider list, which can result in problems when we pass an
ECNamedCurveParameterSpec from a different shaded version of BC.

Fixes: #68

* Add BouncyCastleConfiguration class for handling BC init

This moves the BouncyCastleProvider creation from CryptoAlgorithm to
BouncyCastleConfiguration, which also contains a static field for classes that
must use a pegged version of BouncyCastle. A newInstance reference is included
from CryptoAlgorithm to BouncyCastleConfiguration to force initialization, since
it depends on BouncyCastle being on the SecurityProvider chain.

* Add comments, and use static method to force class loading

This adds documentation for the purpose of the BouncyCastleConfiguration class,
and uses a static method call to force the BouncyCastleConfiguration class to be
loaded.